Fix thumbnail loading
The thumbnails required an 'lv' cookie that was provided when loading the main page. Instead of again hacking this in again, the cookie management was slightly refactored. The 'getBody' request as well as the 'login' request now properly apply the 'Set-Cookie' headers in ExHentaiAuth, which stores them in a separate store. This does mean existing users will be forced to login again. Additionally bumped gradle because the Android Plugin was complaining.